server/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
python-decouple==3.8
psycopg2-binary==2.9.12
django-cors-headers==4.9.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
